refactor(nyz): abstarct ISerialEvaluator and rename serial evaluation implementation
