+++ ding/worker/collector/__init__.py
@@ -4,8 +4,10 @@ from .base_serial_collector import ISerialCollector, create_serial_collector, ge
 from .sample_serial_collector import SampleCollector
 from .episode_serial_collector import EpisodeCollector
 from .episode_one_vs_one_serial_collector import Episode1v1Collector
-from .base_serial_evaluator import BaseSerialEvaluator
-from .one_vs_one_serial_evaluator import OnevOneEvaluator
+
+from .base_serial_evaluator import ISerialEvaluator, VectorEvalMonitor
+from .interaction_serial_evaluator import InteractionSerialEvaluator
+from .battle_interaction_serial_evaluator import BattleInteractionSerialEvaluator
 # parallel
 from .base_parallel_collector import BaseCollector, create_parallel_collector, get_parallel_collector_cls
 from .zergling_collector import ZerglingCollector
